fix: poprawka konfliktów po rebase
